removed common::function in favor of pointers, timers compile again but there's still a deadlock situation when using event dispatchers, removed boost sources, added lost.guiro.view.Button:click()

+++ projects/lostengine/lost/platform/Platform.cpp
@@ -24,13 +24,13 @@ namespace lost
     // returns current time in seconds
     double currentTimeSeconds()
     {
-      return (currentTimeMicroSeconds() / 1000000.0);
+      return ((double) currentTimeMicroSeconds() / 1000000.0);
     }
 
     // returns current time in milliseconds
-    double currentTimeMilliSeconds()
+    long currentTimeMilliSeconds()
     {
-      return (currentTimeMicroSeconds() / 1000.0);
+      return (long) (currentTimeMicroSeconds() / 1000);
     }
 
     PlatformType getPlatform()
