+++ projects/lostengine/lost/platform/Platform_Linux.cpp
@@ -50,12 +50,12 @@ namespace lost
     }
 
     // returns current time in microseconds
-    double currentTimeMicroSeconds()
+    long currentTimeMicroSeconds()
     {
       struct timeval tv;
 
       gettimeofday(&tv, NULL);
-      return ((double)tv.tv_sec)*1000000.0 + (double)tv.tv_usec;
+      return tv.tv_sec*1000000 + tv.tv_usec;
     }
 
     lost::fs::Path getApplicationFilename(bool excludeExtension = false)
