+++ projects/lostengine/lost/platform/Time.h
@@ -22,8 +22,8 @@ namespace lost
 namespace platform
 {
     double currentTimeSeconds(); // returns current time in seconds
-    double currentTimeMilliSeconds(); // returns current time in milliseconds
-    double currentTimeMicroSeconds(); // returns current time in microseconds
+    long currentTimeMilliSeconds(); // returns current time in milliseconds
+    long currentTimeMicroSeconds(); // returns current time in microseconds
     string currentTimeFormat(); // returns current time as formatted string (e.g. "2007/11/26 23:30:37:123")
 }
 }
